feat: remove 128 bit numbers from language

+++ src/typed_ast.rs
@@ -212,9 +212,9 @@ pub enum ExprEnum {
     /// Literal `false`.
     False,
     /// Unsigned number literal.
-    NumUnsigned(u128, UnsignedNumType),
+    NumUnsigned(u64, UnsignedNumType),
     /// Signed number literal.
-    NumSigned(i128, SignedNumType),
+    NumSigned(i64, SignedNumType),
     /// Identifier (either a variable or a function).
     Identifier(String),
     /// Array literal which explicitly specifies all of its elements.
@@ -248,7 +248,7 @@ pub enum ExprEnum {
     /// Explicit cast of an expression to the specified type.
     Cast(Type, Box<Expr>),
     /// Range of numbers from the specified min (inclusive) to the specified max (exclusive).
-    Range((u128, UnsignedNumType), (u128, UnsignedNumType)),
+    Range((u64, UnsignedNumType), (u64, UnsignedNumType)),
 }
 
 /// A variant literal, used by [`ExprEnum::EnumLiteral`], with type and location in the source code.
@@ -367,9 +367,9 @@ pub enum PatternEnum {
     /// Matches `false`.
     False,
     /// Matches the specified unsigned number.
-    NumUnsigned(u128),
+    NumUnsigned(u64),
     /// Matches the specified signed number.
-    NumSigned(i128),
+    NumSigned(i64),
     /// Matches a tuple if all of its fields match their respective patterns.
     Tuple(Vec<Pattern>),
     /// Matches a struct if all of its fields match their respective patterns.
@@ -381,7 +381,7 @@ pub enum PatternEnum {
     /// Matches an enum with the specified name and variant, if all fields match.
     EnumTuple(String, String, Vec<Pattern>),
     /// Matches any number inside the usigned range between min (inclusive) and max (inclusive).
-    UnsignedInclusiveRange(u128, u128),
+    UnsignedInclusiveRange(u64, u64),
     /// Matches any number inside the signed range between min (inclusive) and max (inclusive).
-    SignedInclusiveRange(i128, i128),
+    SignedInclusiveRange(i64, i64),
 }
